+++ src/test/scala/controller/controllerComponent/controllerBaseImpl/ControllerSpec.scala
@@ -1,0 +1,10 @@
+package de.htwg.se.reversi.controller.controllerComponent.controllerBaseImpl
+
+import de.htwg.se.reversi.model.gridComponent.gridBaseImpl.Grid
+
+import scala.language.reflectiveCalls
+import org.scalatest.{Matchers, WordSpec}
+
+class ControllerSpec extends WordSpec with Matchers {
+
+}
