+++ src/test/scala/model/gridComponent/gridBaseImpl/CellSpec.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.reversi.model
+package model.gridComponent.gridBaseImpl
 
 import de.htwg.se.reversi.model.gridComponent.gridBaseImpl.Cell
 import org.scalatest.{Matchers, WordSpec}
